highlighters: Use LineNumberCursor for the separator of number-lines -hlcursor
Use the `LineNumberCursor` face to color the separators `DisplayAtom`
on the the line of the cursor. The local variable `face` always used the
`LineNumbers` face for all separators which looked inconsistent when the
cursor line or wrapped line was using a different face that `LineNumbers`.

This now builds a `Array<DisplayAtom, 2>` on the stack. One element is the
line number, the other is the separator atom. Both atoms are now prepended
to the line within only one `std::vector::insert` call. Since inserting at
the beginning of a vector needs to move all elements this should be about
twice as efficient.